whisper-service/requirements.txt
flask
flask-cors
openai-whisper

